--- app/ports/models/__init__.py
@@ -1,1 +1,0 @@
-from .stats import UUID, PortInstallation, Submission
