+++ app/ports/templates/ports/layout.html
@@ -86,7 +86,7 @@
                 <i><a class="footer-links" href="{% url 'all_builds' %}">{{ latest_build_fetched_at|date:'Y-m-d H:i' }} (UTC)</a></i><br><br>
 
                 Latest stats submission was received at:<br>
-                <i><a class="footer-links" href="{% url 'stats_home' %}">{{ latest_submission_made_at|date:'Y-m-d H:i' }} (UTC)</a></i><br><br>
+                <i><a class="footer-links" href="{% url 'stats' %}">{{ latest_submission_made_at|date:'Y-m-d H:i' }} (UTC)</a></i><br><br>
 
             </div>
     </div>
