+++ app/stats/models.py
