+++ app/stats/templates/stats/port_installations_table.html
