+++ app/stats/templates/stats/stats_port_installations.html
@@ -12,7 +12,7 @@
     <br>
     <ul class="nav nav-tabs" id="tabs" role="tablist">
         <li class="nav-item">
-            <a href="{% url 'stats_home' %}" class="nav-link">System Stats</a>
+            <a href="{% url 'stats' %}" class="nav-link">System Stats</a>
         </li>
         <li class="nav-item">
             <a href="#" class="nav-link active">Port Installations</a>
